Reduce YouTube API calls using .count, not .size
In principle, using `.size` is a good idea. However, `yt-core` is
not currently memoizing `size` so every call to `size` is a new call
to the YouTube API.

Since we are already using `.count` to iterate through all the items,
we are better off using `.count` everywhere so we get the memoized value.